@@ -84,112 +87,94 @@
     "yup": "0.32.x"
   },
   "devDependencies": {
-    "@ant-design-vue/vue-jsx-hot-loader": "0.1.4",
-    "@babel/core": "7.21.0",
-    "@babel/eslint-parser": "7.19.1",
-    "@babel/plugin-proposal-decorators": "7.21.0",
-    "@babel/plugin-transform-runtime": "7.21.0",
-    "@babel/preset-env": "7.20.2",
-    "@babel/preset-typescript": "7.21.0",
-    "@babel/runtime-corejs3": "7.21.0",
-    "@commitlint/cli": "12.1.4",
-    "@commitlint/config-conventional": "12.1.4",
+    "@babel/core": "7.22.0",
+    "@babel/eslint-parser": "7.21.8",
+    "@babel/plugin-proposal-decorators": "7.22.0",
+    "@babel/plugin-transform-class-properties": "^7.22.0",
+    "@babel/plugin-transform-runtime": "7.22.0",
+    "@babel/preset-env": "7.22.0",
+    "@babel/preset-typescript": "7.21.5",
+    "@babel/runtime-corejs3": "7.22.0",
+    "@commitlint/cli": "17.6.3",
+    "@commitlint/config-conventional": "17.6.3",
     "@fatesigner/cz-customizable": "6.3.0",
-    "@semantic-release/changelog": "5.0.1",
-    "@semantic-release/git": "9.0.0",
-    "@tailwindcss/line-clamp": "0.4.2",
-    "@types/chai": "4.3.4",
-    "@types/gulp": "4.0.10",
-    "@types/jest": "27.4.1",
-    "@types/lodash": "4.14.191",
-    "@types/lodash-es": "4.17.6",
+    "@originjs/vite-plugin-commonjs": "1.0.3",
+    "@semantic-release/changelog": "6.0.3",
+    "@semantic-release/git": "10.0.1",
+    "@types/chai": "4.3.5",
+    "@types/gulp": "4.0.11",
+    "@types/jest": "29.5.1",
+    "@types/lodash": "4.14.195",
+    "@types/lodash-es": "4.17.7",
     "@types/mockjs": "1.0.7",
     "@types/node": "16.11.24",
-    "@typescript-eslint/eslint-plugin": "5.53.0",
-    "@typescript-eslint/parser": "5.53.0",
+    "@typescript-eslint/eslint-plugin": "5.59.7",
+    "@typescript-eslint/parser": "5.59.7",
+    "@vitejs/plugin-vue": "4.2.3",
+    "@vitejs/plugin-vue-jsx": "3.0.1",
     "@vue/babel-plugin-jsx": "1.1.1",
     "@vue/babel-preset-jsx": "1.4.0",
-    "@vue/compiler-sfc": "3.2.47",
-    "antd-dayjs-webpack-plugin": "1.0.6",
-    "autoprefixer": "10.4.13",
-    "babel-eslint": "10.1.0",
-    "babel-jest": "27.5.1",
-    "babel-loader": "8.2.3",
+    "@vue/compiler-sfc": "3.3.4",
+    "autoprefixer": "10.4.14",
+    "babel-jest": "29.5.0",
     "babel-plugin-transform-typescript-metadata": "0.3.2",
     "chai": "4.3.7",
-    "chalk": "4.1.2",
+    "chalk": "5.2.0",
     "cheerio": "0.22.0",
+    "cjs2esmodule": "^1.1.1",
     "clean-css": "5.3.2",
     "commitizen": "4.3.0",
-    "copy-webpack-plugin": "8.1.1",
     "cross-env": "7.0.3",
     "cross-spawn": "7.0.3",
-    "css-loader": "4.3.0",
-    "css-minimizer-webpack-plugin": "1.3.0",
-    "dotenv": "10.0.0",
-    "dotenv-expand": "9.0.0",
-    "ejs": "3.1.8",
-    "eslint": "8.35.0",
-    "eslint-config-prettier": "8.6.0",
+    "dotenv": "16.0.3",
+    "dotenv-expand": "10.0.0",
+    "ejs": "3.1.9",
+    "eslint": "8.41.0",
+    "eslint-config-prettier": "8.8.0",
     "eslint-config-standard": "17.0.0",
-    "eslint-plugin-html": "6.2.0",
+    "eslint-define-config": "1.20.0",
+    "eslint-plugin-html": "7.1.0",
     "eslint-plugin-import": "2.27.5",
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prettier": "4.2.1",
     "eslint-plugin-promise": "6.1.1",
     "eslint-plugin-simple-import-sort": "10.0.0",
+    "eslint-plugin-vue": "9.14.1",
     "exit": "0.1.2",
-    "file-loader": "6.2.0",
-    "fork-ts-checker-webpack-plugin": "6.5.0",
-    "fs-extra": "9.1.0",
-    "globby": "11.1.0",
+    "fs-extra": "11.1.1",
+    "globby": "13.1.4",
     "gulp": "4.0.2",
-    "happypack": "5.0.1",
-    "html-webpack-plugin": "5.5.0",
-    "husky": "7.0.4",
-    "jest": "27.5.1",
+    "husky": "8.0.3",
+    "jest": "29.5.0",
     "less": "4.1.3",
-    "less-loader": "10.2.0",
-    "lint-staged": "11.2.6",
+    "lint-staged": "13.2.2",
     "lodash": "4.17.21",
-    "mini-css-extract-plugin": "2.7.2",
     "mockjs": "1.1.0",
-    "nprogress": "0.2.0",
-    "object-hash": "2.2.0",
+    "object-hash": "3.0.0",
     "portfinder": "1.0.32",
-    "postcss": "8.4.21",
-    "postcss-loader": "5.3.0",
-    "prettier": "2.8.4",
-    "prettier-plugin-tailwindcss": "0.2.3",
-    "raw-loader": "4.0.2",
+    "postcss": "8.4.23",
+    "prettier": "2.8.8",
+    "prettier-plugin-tailwindcss": "0.3.0",
     "require-dir": "1.2.0",
-    "rimraf": "3.0.2",
-    "sass": "1.58.3",
-    "sass-loader": "11.1.1",
-    "semantic-release": "17.4.4",
+    "rimraf": "5.0.1",
+    "semantic-release": "21.0.2",
     "short-uuid": "4.2.2",
-    "string-width": "2.1.1",
-    "strip-ansi": "5.2.0",
-    "style-loader": "2.0.0",
-    "stylelint": "14.9.1",
+    "string-width": "6.1.0",
+    "strip-ansi": "7.0.1",
+    "stylelint": "15.6.2",
     "stylelint-config-css-modules": "4.2.0",
-    "stylelint-config-recess-order": "3.1.0",
-    "stylelint-config-standard": "26.0.0",
-    "stylelint-order": "5.0.0",
-    "tailwindcss": "3.2.7",
-    "terser": "5.16.5",
-    "terser-webpack-plugin": "5.3.6",
-    "ts-jest": "27.1.3",
-    "ts-loader": "9.4.2",
+    "stylelint-config-recess-order": "4.0.0",
+    "stylelint-config-standard": "33.0.0",
+    "stylelint-config-standard-scss": "9.0.0",
+    "stylelint-order": "6.0.3",
+    "tailwindcss": "3.3.2",
+    "ts-jest": "29.1.0",
     "ts-node": "10.9.1",
-    "typescript": "4.9.5",
-    "url-loader": "4.1.1",
-    "vue-loader": "16.8.3",
-    "webpack": "5.75.0",
-    "webpack-bundle-analyzer": "4.8.0",
-    "webpack-cli": "4.10.0",
-    "webpack-dev-server": "4.11.1",
-    "webpack-manifest-plugin": "4.1.1"
+    "typescript": "5.0.4",
+    "vite": "4.3.9",
+    "vite-plugin-ejs": "1.6.4",
+    "vite-plugin-importer": "0.2.5",
+    "vue-eslint-parser": "9.3.0"
   },
   "config": {
     "commitizen": {
