@@ -1,14 +1,3 @@
-export * from './types';
 export * from './button';
 export * from './button-add';
-export * from './button-delete';
-export * from './button-download';
-export * from './button-edit';
-export * from './button-export';
-export * from './button-fullscreen';
-export * from './button-fullscreen-exit';
-export * from './button-refresh';
-export * from './button-save';
-export * from './button-search';
-export * from './button-spin';
-export * from './button-upload';
+export * from './types';
