@@ -2,7 +2,7 @@ import { defineComponent } from 'vue';
 
 import $styles from './footer.module.less';
 
-export const AppFooter = defineComponent({
+export default defineComponent({
   name: 'AppFooter',
   props: {
     pure: {
@@ -12,9 +12,9 @@ export const AppFooter = defineComponent({
   },
   render(ctx) {
     return (
-      <footer class={['tw-text-center tw-space-y-1', ctx.pure ? undefined : $styles.footer]}>
+      <footer class={['tw-space-y-1 tw-text-center', ctx.pure ? undefined : $styles.footer]}>
         <div class={ctx.pure ? undefined : 'tw-text-xxs tw-text-gray-700'}>Copyright 2021 All Rights Reserved Vue</div>
-        <div class={ctx.pure ? undefined : 'tw-text-xxs tw-text-gray-700'}>(Only Edge、Chrome )</div>
+        <div class={ctx.pure ? undefined : 'tw-text-xxs tw-text-gray-700'}>(Only Edge, Chrome )</div>
       </footer>
     );
   }
