@@ -1,14 +1,14 @@
-import { Subscription } from 'rxjs';
-import { useRoute, useRouter } from 'vue-router';
 import { defineComponent, onMounted, onUnmounted, ref } from 'vue';
+import { useRoute, useRouter } from 'vue-router';
+import { DownOutlined, LockOutlined, LogoutOutlined } from '@ant-design/icons-vue';
 import { Dropdown, Menu, MenuDivider, MenuItem } from 'ant-design-vue';
-import { IconArrowDownSLine, IconLockLine, IconLogoutBoxLine } from 'antdvx';
+import { Subscription } from 'rxjs';
 
-import { i18nMessages } from '@/app/i18n';
-import { AppStore } from '@/app/core/store';
-import { UserType } from '@/app/core/types';
 import { login$, logout$ } from '@/app/core/events';
 import { authService, sessionService } from '@/app/core/services';
+import { AppStore } from '@/app/core/store';
+import { UserType } from '@/app/core/types';
+import { i18nMessages } from '@/app/i18n';
 
 import $styles from './user.module.less';
 
@@ -29,7 +29,7 @@ export const NavUser = defineComponent({
 
     // 个性化弹出层
     const individuation = () => {
-      AppStore.getPopupRefs().individuation.handler.present();
+      // AppStore.getPopupRefs().individuation.handler.present();
     };
 
     // 切换角色
@@ -106,11 +106,10 @@ export const NavUser = defineComponent({
                 </Menu>
               ) : undefined;
             }
-          }}
-        >
+          }}>
           <div class={$styles.dropdown}>
             <div class={$styles.username}>{ctx.currentRole?.[0]}</div>
-            <IconArrowDownSLine />
+            <DownOutlined />
           </div>
         </Dropdown>
         <Dropdown
@@ -128,22 +127,21 @@ export const NavUser = defineComponent({
                   <MenuDivider /> */}
                   <MenuItem onClick={ctx.updatePassword}>
                     <div class='tw-flex tw-items-center tw-gap-1'>
-                      <IconLockLine />
+                      <LockOutlined />
                       <span>{ctx.$t(i18nMessages.app.navbar.updatePassword)}</span>
                     </div>
                   </MenuItem>
                   <MenuDivider />
                   <MenuItem onClick={ctx.logout}>
                     <div class='tw-flex tw-items-center tw-gap-1'>
-                      <IconLogoutBoxLine />
+                      <LogoutOutlined />
                       <span>{ctx.$t(i18nMessages.app.navbar.logOut)}</span>
                     </div>
                   </MenuItem>
                 </Menu>
               );
             }
-          }}
-        >
+          }}>
           <div class={$styles.dropdown}>
             <div class={$styles.username}>{ctx.user.username}</div>
             <img class={$styles.avatar} src={ctx.user.avatar} alt='' title='' />
