@@ -2,19 +2,8 @@
  * Antdvx
  */
 
-import { Empty, message, notification } from 'ant-design-vue';
-import {
-  configureEcharts,
-  configureVEcharts,
-  configureXTable,
-  setRequestAdapter,
-  setStorageService,
-  XButtonExport,
-  XButtonFullscreen,
-  XButtonFullscreenExit,
-  XButtonRefresh,
-  XTableSettingsPanelButton
-} from 'antdvx';
+import { message, notification } from 'ant-design-vue';
+import { setRequestAdapter, setStorageService } from 'antdvx';
 import { setAntdvxPipesConfig } from 'antdvx/pipes';
 import { merge } from 'lodash-es';
 
@@ -61,55 +50,6 @@ export const Antdvx = {
     const classes = document.body.className.split(' ');
     classes.push('antdvx-theme-classic');
     document.body.className = classes.filter((x: any) => !!x).join(' ');
-
-    // echarts
-    configureEcharts({
-      color: ['#91cc75', '#fac858', '#5470c6', '#ee6666', '#73c0de', '#3ba272', '#fc8452', '#9a60b4', '#ea7ccc'],
-      textStyle: {
-        fontFamily: `-apple-system, BlinkMacSystemFont, Roboto, 'Noto Sans', 'Liberation Sans', 'PingFang SC', 'Hiragino Sans GB', 'Heiti SC', 'Microsoft Jhenghei', 'WenQuanYi Micro Hei', sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji'`
-      }
-    });
-    configureVEcharts({
-      header(chartRef) {
-        return (
-          <div class='tw-mb-2 tw-flex tw-items-center tw-justify-end'>
-            <div class='tw-flex-initial'>
-              <img width='40' height='20' src={require('@/assets/img/logo.png')} alt='' title='' />
-            </div>
-            {chartRef.title ? (
-              <div class='tw-flex-1 tw-pl-2 tw-pr-2'>
-                <div class='tw-text-center tw-text-base'>{chartRef.title}</div>
-              </div>
-            ) : (
-              ''
-            )}
-            {chartRef.error ? (
-              ''
-            ) : (
-              <div class='tw-flex-initial tw-space-x-2'>
-                {chartRef.refreshable ? (
-                  <XButtonRefresh
-                    disabled={chartRef.loading}
-                    color='primary'
-                    size='small'
-                    type='link'
-                    handler={chartRef.refresh}
-                  />
-                ) : undefined}
-                {chartRef.exportable ? (
-                  <XButtonExport
-                    disabled={chartRef.loading}
-                    size='small'
-                    placement='bottomRight'
-                    options={chartRef.exportOptions}
-                  />
-                ) : undefined}
-              </div>
-            )}
-          </div>
-        );
-      }
-    });
 
     const loadLang = async (lang) => {
       // 导入 language，非中文环境统一使用英文
