@@ -1,6 +1,7 @@
 import { defineComponent, onDeactivated, PropType, reactive, ref } from 'vue';
+import { LockOutlined, UserOutlined } from '@ant-design/icons-vue';
 import { Alert, Form, FormItem, Input, InputPassword, notification } from 'ant-design-vue';
-import { IconLockLine, IconLockUnlockLine, IconUserLine, SlideCaptcha, TransitionCollapse, XButton } from 'antdvx';
+import { SlideCaptcha, TransitionCollapse, XButton } from 'antdvx';
 import to from 'await-to-js';
 import { Field as VeeField } from 'vee-validate';
 
@@ -146,7 +147,7 @@ export default defineComponent({
                       }}
                       v-slots={{
                         prefix() {
-                          return <IconUserLine />;
+                          return <UserOutlined />;
                         }
                       }}
                     />,
@@ -179,7 +180,7 @@ export default defineComponent({
                     }}
                     v-slots={{
                       prefix() {
-                        return <IconLockLine />;
+                        return <LockOutlined />;
                       }
                     }}
                   />,
@@ -211,7 +212,7 @@ export default defineComponent({
                     }}
                     v-slots={{
                       prefix() {
-                        return <IconLockUnlockLine />;
+                        return <LockOutlined />;
                       }
                     }}
                   />,
@@ -243,7 +244,7 @@ export default defineComponent({
                     }}
                     v-slots={{
                       prefix() {
-                        return <IconLockUnlockLine />;
+                        return <LockOutlined />;
                       }
                     }}
                   />,
@@ -283,8 +284,7 @@ export default defineComponent({
             size='large'
             type='primary'
             loading={ctx.form.isSubmitting}
-            onClick={ctx.form.submit}
-          >
+            onClick={ctx.form.submit}>
             {ctx.$t(i18nMessages.app.passport.updatePassword.submit)}
           </XButton>
           {ctx.$slots?.extra?.()}
