@@ -7,8 +7,6 @@ import nprogress from 'nprogress';
 
 // App
 import { App } from '@/app/app';
-import { Directives } from '@/app/core/directives';
-import { Pipes } from '@/app/core/pipes';
 // Plugins
 import { i18n } from '@/app/i18n';
 import { Antdvx } from '@/app/plugins/antdvx';
@@ -25,7 +23,7 @@ async function mountApp() {
   const app = createApp(App);
   const router = await createAppRouter();
 
-  app.use(router).use(i18n._).use(Antdvx).use(Dayjs).use(Exceljs).use(VeeValidate).use(Pipes).use(Directives);
+  app.use(router).use(i18n._).use(Antdvx).use(Dayjs).use(Exceljs).use(VeeValidate);
 
   router.isReady().then(() => {
     // Nprogress 延迟 100ms 显示
